+++ src/pages/RegisterPage.tsx
@@ -1,4 +1,4 @@
-import Radio from 'components/atoms/Radio';
+import Radio from 'components/molecules/StatusSelectGroup';
 import DayModalInput from 'components/molecules/DayModalInput';
 import MRegisterForm from 'components/organisms/MRegisterForm';
 import RegisterHeader from 'components/organisms/RegisterHeader';
@@ -7,11 +7,11 @@ import GNB from 'layouts/GNB';
 const RegisterPage = () => {
   return (
     <div>
-      <GNB />
+      {/* <GNB />
       <RegisterHeader />
-      <MRegisterForm />
+      <MRegisterForm /> */}
       <DayModalInput />
-      <Radio />
+      {/* <Radio /> */}
     </div>
   );
 };
